requirements.txt
python-dotenv
google-genai
google-api-core
openai>=1.60.0
Pillow
streamlit
reportlab
ebooklib
pymupdf  # PDF page extraction for sample preview

# Database
sqlalchemy>=2.0
psycopg2-binary
alembic

# Auth (password hashing)
passlib[bcrypt]
bcrypt>=4.0,<4.1  # Pin to avoid passlib compatibility issue with 4.1+

# Storage
boto3

# Encryption (for API key storage)
cryptography
